Ping server before test
